+++ tools/benchmark/CHANGELOG.md
@@ -6,6 +6,8 @@
 -   Geometric mean has been restored to the results summary (was removed in the past) and correctly handles both larger-is-better and smaller-is-better values. This should make evaluating overall impact of changes easier.
 -   All benchmarks now have a primary result (included in the geometric mean), and optional additional results.
 -   Formatting of console output from reporter has changed.
+-   Suites with colliding names now emit a warning and a disambiguation number is added as a suffix.
+-	Multiple benchmarks with the same name in the same suite no longer overwrite results, and emit a warning.
 
 ### ⚠ BREAKING CHANGES